+++ app/src/main/java/com/hbbsolution/owner/work_management/view/payment/view/PaymentView.java
@@ -7,7 +7,7 @@ import com.hbbsolution.owner.work_management.model.billGv24.BillGv24Response;
  */
 
 public interface PaymentView {
-    void getWalletSuccess(int wallet);
+    void getWalletSuccess(long wallet);
     void getWalletFail();
     void getInfoBill24h(BillGv24Response billGv24Response);
     void getErrorBill24h(String error);
